+++ src/DependencyInjection/CompilerPass/ObjectMapperPass.php
@@ -35,18 +35,19 @@ final readonly class ObjectMapperPass implements CompilerPassInterface
             $serviceDefinition = $container->getDefinition($serviceId);
             $serviceClass = $serviceDefinition->getClass() ?? throw new InvalidArgumentException('Class is required');
 
-            /** @var array<string,string> $tag */
+            /** @var array{sourceClass:class-string,targetClass:class-string,serviceId:string,method:string,hasExistingTarget:bool,extraArguments:array<string,int>} $tag */
             foreach ($tags as $tag) {
                 $method = $tag['method'] ?? throw new InvalidArgumentException('Method is required');
 
                 $objectMapperTableFactory->addMethodCall(
                     'addObjectMapper',
                     [
-                        $tag['sourceClass'],
-                        $tag['targetClass'],
-                        $serviceId,
-                        $method,
-                        ServiceMethodExtraArgumentUtil::getExtraArguments($serviceClass, $method),
+                        '$sourceClass' => $tag['sourceClass'],
+                        '$targetClass' => $tag['targetClass'],
+                        '$serviceId' => $serviceId,
+                        '$method' => $method,
+                        '$hasExistingTarget' => $tag['hasExistingTarget'],
+                        '$extraArguments' => ServiceMethodExtraArgumentUtil::getExtraArguments($serviceClass, $method, $tag['hasExistingTarget']),
                     ],
                 );
             }