+++ src/Transformer/Implementation/ObjectMapperTransformer.php
@@ -85,9 +85,10 @@ final class ObjectMapperTransformer implements
             subMapperFactory: $this->subMapperFactory,
         );
 
-        return $serviceMethodRunner->run(
+        return $serviceMethodRunner->runObjectMapper(
             serviceMethodSpecification: $serviceMethodSpecification,
             source: $source,
+            target: $target,
             targetType: $targetType,
             context: $context,
         );